Preserve original httplib response. (#7)
* Setup a manual http server. Add unittest for multiple Set-Cookie header.

* Add cherrypy based test for set-cookie through redirects.

* Use source files for -dev tox environment.

* Preserve original httplib response.

Requests expects a low-level response object at
Response.raw._original_response for its cookie handling. This is currently lost
in the interface between VTS and responses. This PR stashes the response and
retrieves it later to permit the proper behaviour of cookies.

* Bump version and pin urllib3

Urllib3 is failing above version 1.25 with a certificate issue. Pin it
for the time being.

See https://github.com/urllib3/urllib3/issues/1603

* Fix cookie-handling to cope with multiple cookies.

- Vendor in and patch responses to handle cookies sent in
comma-separated headers
- Update tests